refactor: 강퇴된 사용자에게 알림창 출력 여부 확인 로직 훅으로 분리
