+++ src/app/(main)/_components/molecules/AgoraListDecider.tsx
@@ -8,10 +8,10 @@ import { AgoraTabStatus, SearchParams } from '@/app/model/Agora';
 import Loading from '@/app/_components/atoms/loading';
 import { useSearchStore } from '@/store/search';
 import { useShallow } from 'zustand/react/shallow';
-import { useKickedStore } from '@/store/kick';
-import { swalKickedUserAlert } from '@/utils/swalAlert';
+import { useAlertKickedUser } from '@/hooks/useAlertKickedUser';
 import LivelyAgoraList from './LivelyAgoraList';
 import CategoryAgoraNowTitle from '../atoms/CategoryAgoraNowTitle';
+import Borderline from '../../user-info/_component/atoms/Borderline';
 
 const KeywordAgoraList = dynamic(() => import('./KeywordAgoraList'), {
   loading: () => (
@@ -47,32 +47,13 @@ export default function AgoraListDecider({ searchParams }: Props) {
     })),
   );
 
-  const { kicked, reset } = useKickedStore(
-    useShallow((state) => ({
-      kicked: state.kicked,
-      reset: state.reset,
-    })),
-  );
-
   useEffect(() => {
     if (q) {
       setSearch(q);
     }
   }, [q, setSearch]);
 
-  useEffect(() => {
-    const handleKicked = async () => {
-      if (kicked) {
-        const result = await swalKickedUserAlert();
-
-        if (result && result.isConfirmed) {
-          reset();
-        }
-      }
-    };
-
-    handleKicked();
-  }, [kicked]);
+  useAlertKickedUser();
 
   if (search) {
     return (
@@ -89,9 +70,7 @@ export default function AgoraListDecider({ searchParams }: Props) {
           <ErrorBoundary FallbackComponent={FallbackComponent}>
             <LivelyAgoraList />
           </ErrorBoundary>
-          <div className="h-6 w-full mb-16">
-            <div className="h-full mx-10 bg-gray-400 opacity-15" />
-          </div>
+          <Borderline className="mb-16 mx-10" />
         </>
       )}
       <ErrorBoundary FallbackComponent={FallbackComponent}>
